Pull out logic for infographics into a util
